Rename spotify package
