+++ src/packages/music_api_clients/models/artist.py
