+++ src/packages/music_api_clients/models/playlist.py
@@ -1,4 +1,4 @@
-from packages.spotify.models.track import Track
+from packages.music_api_clients.models.track import Track
 
 class Playlist:
     def __init__(self, id_, name, description, tracks_fetcher):
