+++ src/packages/music_api_clients/models/song_attribute_ranges.py
@@ -1,4 +1,4 @@
-from packages.spotify.models.audio_features import AudioFeatures
+from packages.music_api_clients.models.audio_features import AudioFeatures
 
 
 class SongAttributeRanges:
