+++ src/packages/music_api_clients/models/track.py
@@ -1,4 +1,4 @@
-from packages.spotify.models.artist import Artist
+from packages.music_api_clients.models.artist import Artist
 
 
 class Track:
