+++ src/scripts/get_playlist_metadata.py
@@ -6,7 +6,7 @@ sys.path.extend(['.', '../'])
 
 from packages.music_management.music_util import MusicUtil
 from packages.music_management.my_music_lib import MyMusicLib
-from packages.spotify.spotify_client_wrapper import SpotifyClientWrapper
+from packages.music_api_clients.spotify_client_wrapper import SpotifyClientWrapper
 
 def get_playlist_metadata(my_music_lib, playlist_name):
     playlist = my_music_lib.get_playlist_by_name(playlist_name)
