+++ src/scripts/update_seed_playlists.py
@@ -5,7 +5,7 @@ sys.path.extend(['.', '../'])
 
 from packages.music_management.my_music_lib import MyMusicLib
 from packages.music_management.music_util import MusicUtil
-from packages.spotify.spotify_client_wrapper import SpotifyClientWrapper
+from packages.music_api_clients.spotify_client_wrapper import SpotifyClientWrapper
 
 NUM_TRACKS_PER_ALBUM = 3
